+++ app/preact/src/server/framework-preset-preact.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 import { TransformOptions } from '@babel/core';
 import { Configuration } from 'webpack';
+import { findDistEsm, StorybookConfig } from '@storybook/core-common';
 
 export function babelDefault(config: TransformOptions) {
   return {
@@ -26,3 +27,7 @@ export function webpackFinal(config: Configuration) {
     },
   };
 }
+
+export const config: StorybookConfig['config'] = (entry = []) => {
+  return [...entry, findDistEsm(__dirname, 'client/preview/config')];
+};