+++ app/react/src/server/options.ts
@@ -4,9 +4,5 @@ import { LoadOptions } from '@storybook/core-common';
 export default {
   packageJson: sync({ cwd: __dirname }).packageJson,
   framework: 'react',
-  frameworkPresets: [
-    require.resolve('./framework-preset-react'),
-    require.resolve('./framework-preset-cra'),
-    require.resolve('./framework-preset-react-docgen'),
-  ],
+  frameworkPresets: [require.resolve('./preset')],
 } as LoadOptions;