+++ app/vue3/src/server/framework-preset-vue3.ts
@@ -1,5 +1,6 @@
 import { VueLoaderPlugin } from 'vue-loader';
 import { Configuration, DefinePlugin } from 'webpack';
+import { findDistEsm, StorybookConfig } from '@storybook/core-common';
 
 export function webpack(config: Configuration): Configuration {
   return {
@@ -45,3 +46,7 @@ export function webpack(config: Configuration): Configuration {
     },
   };
 }
+
+export const config: StorybookConfig['config'] = (entry = []) => {
+  return [...entry, findDistEsm(__dirname, 'client/preview/config')];
+};