+++ lib/store/src/prepareStory.ts
@@ -28,7 +28,7 @@ const argTypeDefaultValueWarning = deprecate(
   dedent`
   \`argType.defaultValue\` is deprecated and will be removed in Storybook 7.0.
 
-  https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#deprecated-argtype-defaultValue`
+  https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#no-longer-inferring-default-values-of-args`
 );
 
 // Combine all the metadata about a story (both direct and inherited from the component/global scope)
@@ -92,11 +92,11 @@ export function prepareStory<TFramework extends AnyFramework>(
   parameters.__isArgsStory = passArgsFirst && render.length > 0;
 
   // Pull out args[X] into initialArgs for argTypes enhancers
-  const passedArgs: Args = combineParameters(
-    projectAnnotations.args,
-    componentAnnotations.args,
-    storyAnnotations.args
-  ) as Args;
+  const passedArgs: Args = {
+    ...projectAnnotations.args,
+    ...componentAnnotations.args,
+    ...storyAnnotations.args,
+  } as Args;
 
   const contextForEnhancers: StoryContextForEnhancers<TFramework> = {
     componentId: componentAnnotations.id,